Move Redis protocol functionality to a separate package.
