+++ lib/srv/db/redis/cmds.go
@@ -21,7 +21,6 @@ package redis
 import (
 	"context"
 	"fmt"
-	"strconv"
 	"strings"
 
 	"github.com/go-redis/redis/v8"
@@ -209,145 +208,4 @@ func (e *Engine) processAuth(ctx context.Context, redisClient redis.UniversalCli
 		// Redis returns "syntax error" if AUTH has more than 2 arguments.
 		return trace.BadParameter("syntax error")
 	}
-}
-
-// writeCmd writes Redis commands passed as vals to Redis wire form.
-// Most types is covered by go-redis implemented WriteArg() function. Types override by this function are:
-// * Redis errors and Go error: go-redis returns a "human-readable" string instead of RESP compatible error message
-// * integers: go-redis converts them to string, which is not always what we want to.
-// * slices: arrays are recursively converted to RESP responses.
-func writeCmd(wr *redis.Writer, vals interface{}) error {
-	switch val := vals.(type) {
-	case redis.Error:
-		if val == redis.Nil {
-			// go-redis returns nil value as errors, but Redis Wire protocol decodes them differently.
-			// Note: RESP3 has different sequence for nil, current nil is RESP2 compatible as the rest
-			// of our implementation.
-			if _, err := wr.WriteString("$-1\r\n"); err != nil {
-				return trace.Wrap(err)
-			}
-			return nil
-		}
-
-		if err := writeError(wr, "-", val); err != nil {
-			return trace.Wrap(err)
-		}
-	case trace.Error:
-		if err := writeError(wr, "-ERR Teleport: ", val); err != nil {
-			return trace.Wrap(err)
-		}
-	case error:
-		if err := writeError(wr, "-ERR ", val); err != nil {
-			return trace.Wrap(err)
-		}
-	case int:
-		if err := writeInteger(wr, int64(val)); err != nil {
-			return trace.Wrap(err)
-		}
-	case int8:
-		if err := writeInteger(wr, int64(val)); err != nil {
-			return trace.Wrap(err)
-		}
-	case int16:
-		if err := writeInteger(wr, int64(val)); err != nil {
-			return trace.Wrap(err)
-		}
-	case int32:
-		if err := writeInteger(wr, int64(val)); err != nil {
-			return trace.Wrap(err)
-		}
-	case int64:
-		if err := writeInteger(wr, val); err != nil {
-			return trace.Wrap(err)
-		}
-	case []string:
-		if err := writeStringSlice(wr, val); err != nil {
-			return trace.Wrap(err)
-		}
-	case []interface{}:
-		if err := writeSlice(wr, val); err != nil {
-			return trace.Wrap(err)
-		}
-	case interface{}:
-		err := wr.WriteArg(val)
-		if err != nil {
-			return trace.Wrap(err)
-		}
-	}
-
-	return nil
-}
-
-// writeError converts Go error to Redis wire form.
-func writeError(wr *redis.Writer, prefix string, val error) error {
-	if _, err := wr.WriteString(prefix); err != nil {
-		// Add error header specified in https://redis.io/topics/protocol#resp-errors
-		// to follow the convention.
-		return trace.Wrap(err)
-	}
-
-	if _, err := wr.WriteString(val.Error()); err != nil {
-		return trace.Wrap(err)
-	}
-
-	if _, err := wr.Write([]byte("\r\n")); err != nil {
-		return trace.Wrap(err)
-	}
-
-	return nil
-}
-
-// writeSlice converts []interface{} to Redis wire form.
-func writeSlice(wr *redis.Writer, vals []interface{}) error {
-	if err := wr.WriteByte(redis.ArrayReply); err != nil {
-		return trace.Wrap(err)
-	}
-	n := len(vals)
-	if err := wr.WriteLen(n); err != nil {
-		return trace.Wrap(err)
-	}
-
-	for _, v0 := range vals {
-		if err := writeCmd(wr, v0); err != nil {
-			return trace.Wrap(err)
-		}
-	}
-
-	return nil
-}
-
-// writeStringSlice converts []string to Redis wire form.
-func writeStringSlice(wr *redis.Writer, vals []string) error {
-	if err := wr.WriteByte(redis.ArrayReply); err != nil {
-		return trace.Wrap(err)
-	}
-	n := len(vals)
-	if err := wr.WriteLen(n); err != nil {
-		return trace.Wrap(err)
-	}
-
-	for _, v0 := range vals {
-		if err := writeCmd(wr, v0); err != nil {
-			return trace.Wrap(err)
-		}
-	}
-
-	return nil
-}
-
-// writeInteger converts integer to Redis wire form.
-func writeInteger(wr *redis.Writer, val int64) error {
-	if err := wr.WriteByte(redis.IntReply); err != nil {
-		return trace.Wrap(err)
-	}
-
-	v := strconv.FormatInt(val, 10)
-	if _, err := wr.WriteString(v); err != nil {
-		return trace.Wrap(err)
-	}
-
-	if _, err := wr.Write([]byte("\r\n")); err != nil {
-		return trace.Wrap(err)
-	}
-	return nil
 }
